Fix issue where wrong command is executed for input keys
During input-decode-map processing, the current buffer is the keyboard
buffer. This causes the command to be determined by the keyboard
buffer's local keymap, and that command is then executed in the
original buffer. For example, RET in shell-mode or C-j in
lisp-interaction-mode do not work as expected, and only the newline
command is executed instead.